Address PR review comments
- Split factory-test-realm.ts into cohesive modules:
  test-run-types.ts, test-run-parsing.ts, test-run-cards.ts,
  test-run-execution.ts (barrel re-export in factory-test-realm.ts)
- Make passedCount/failedCount computed fields on TestRun card
  (derived from results array, not stored)
- Fix double-pull: pull target realm once, find specs in same dir
- Remove unused darkfactoryModuleUrl param from ensureTestArtifactsRealm
- Fix findSpecFiles doc comment to match actual behavior (no fallback)
- Propagate write failure in ensureTestArtifactsRealm
- Revert src/ → scripts/lib/ import for cardSourceMimeType (keep
  local constants in factory-brief.ts and factory-bootstrap.ts)
- Make combined cache:prepare the default for multiple realms
  (remove --combined flag requirement)
- Fix formatting (prettier)
