+++ packages/software-factory/src/factory-brief.ts
@@ -5,7 +5,8 @@ const wikiLinkPattern = /\[\[([^[\]]+)\]\]/g;
 const markdownHeadingReplacePattern = /^\s*#{1,6}\s+/gm;
 const markdownListReplacePattern = /^\s*[-*+]\s+/gm;
 const whitespacePattern = /\s+/g;
-import { cardSourceMimeType } from '../scripts/lib/realm-operations';
+// Matches SupportedMimeType.CardSource from @cardstack/runtime-common (see CS-10550).
+const cardSourceMimeType = 'application/vnd.card+source';
 
 export interface FactoryBrief {
   title: string;
